+++ gcs/CHANGES.md
@@ -1,5 +1,7 @@
 # Release Notes
 
+## Next
+
 ## 3.0.6 - 2025-04-02
 1. Add AUTO_RANDOM as new fadvise mode.